requirements.txt
streamlit==1.29.0
google-generativeai==0.3.0
requests==2.31.0
python-dotenv==1.0.0
folium==0.14.0
streamlit-folium==0.15.1
pydeck==0.8.1b0
geopy==2.3.0
SpeechRecognition==3.10.0
pydub==0.25.1
googletrans==4.0.0-rc1
google-maps-services-python==4.10.0
timezonefinder==6.2.0
timezonefinder